+++ modules/client/src/tokenVoting/interfaces.ts
@@ -123,6 +123,7 @@ export type TokenVotingProposal = ProposalBase & {
 export type TokenVotingProposalListItem = ProposalListItemBase & {
   token: Erc20TokenDetails | Erc721TokenDetails | null;
   result: TokenVotingProposalResult;
+  totalVotingWeight: bigint;
 };
 
 export type TokenVotingProposalResult = {
@@ -153,6 +154,7 @@ export type SubgraphTokenVotingProposalListItem = SubgraphProposalBase & {
   plugin: {
     token: SubgraphErc20Token | SubgraphErc721Token;
   };
+  totalVotingPower: string;
 };
 
 type SubgraphBaseToken = {