+++ modules/client/CHANGELOG.md
@@ -24,6 +24,7 @@ TEMPLATE:
 - JsDoc for new encoders and decoders
 - Add `settings` and `appovals` fields to `MultisigProposalBase`
 - Add `creationBlockNumber` `executionDate` and `creationBlockNumber` to `MultisigProposal`
+- Add `totalVotingWeight` to `tokenVotingListItem`
 ### Changed
 - `ICreateProposalParams` => `CreateMajorityVotingProposalParams`
 - `IProposalSettings` => `MajorityVotingProposalSettings`